upgrade to latest

+++ src/game.cpp
@@ -39,7 +39,7 @@ xx::Task<> Game::Render2Logic_() {
 void Game::Render2() {
     Render2Logic();
     ShaderSwitch(shader_IndexTriangles);
-    auto [vs, is] = shader_IndexTriangles.Alloc(&cb, std::size(vertices2), std::size(indices2));
+    auto [vs, is] = shader_IndexTriangles.Alloc(&cb, (UINT)std::size(vertices2), (UINT)std::size(indices2));
     memcpy(vs, vertices2, sizeof(vertices2));
     memcpy(is, indices2, sizeof(indices2));
 }
